chore: move getInitials function to separate file
